requirements.txt
schemathesis==4.0.2
allure-pytest>=2.14.3